rm cast() from Property class
* Fixed Property.getValueType() to return a properly ’T’ typed ValueType
* ValueType, which is exposed by Property, was moved to the main
package, since it is a public API
