+++ andhow/src/main/java/yarnandtail/andhow/valuetype/BaseValueType.java
@@ -1,5 +1,6 @@
 package yarnandtail.andhow.valuetype;
 
+import yarnandtail.andhow.ValueType;
 import yarnandtail.andhow.TextUtil;
 import yarnandtail.andhow.load.ParsingException;
